Map: Extract lit-completeness bookmark migration
- Old Vollständigkeit checkbox still lands on the roads completeness style
- Parking ID maps stay as they were on develop
